refactor: rename internal/taskfileserver to internal/server
Pure rename of the taskfileserver package to server, with no behaviour
changes. Files preserve history via git mv; package declarations and
import paths are updated to match.

Refs #80.
