+++ internal/server/doc.go
@@ -1,0 +1,2 @@
+// Package server implements the Taskfile-backed MCP server.
+package server
