+++ internal/server/logging_test.go
@@ -1,4 +1,4 @@
-package taskfileserver
+package server
 
 import (
 	"bytes"
