+++ internal/server/server.go
@@ -1,4 +1,4 @@
-package taskfileserver
+package server
 
 import (
 	"context"
