+++ internal/server/state.go
@@ -1,4 +1,4 @@
-package taskfileserver
+package server
 
 import (
 	"log/slog"
