--- internal/taskfileserver/doc.go
@@ -1,2 +1,0 @@
-// Package taskfileserver implements the Taskfile-backed MCP server.
-package taskfileserver
